Debugging the parallelism part
Added some print statements to help.
Currently there is a segmentation fault when applying the iterators. You
may see the limits of each process/subdomain and which rank fails on
which operation.

Probably the problem is at the offsets or the sizes of the buffers.

Please note that in the current setup, each process has one ghost layer
in each dimension (e.g. only to the right and top, or only to the left
and top etc.).