Migrated to non-blocking communication.
The MPI_Sendrecv() calls were replaced with combinations of MPI_Isend(),
MPI_Irecv(), MPI_Wait().

Each process, e.g. in the communicatePressure():
1. Fills the Left and Right SEND buffers.
2. Posts send and receive requests for the L-->R and L<--R.
3. Waits for its receives requests (only) to complete.
4. Reads the Left and Right RECEIVE buffers.
5. Without waiting for anything else, fills the Bottom and Top SEND
buffers. These are different buffers than the previous.
6. Repeats the steps 2 to 4 for B-->T and B<--T.
7. Without waiting for anything else, fills the Front and Back SEND
buffers.
8. Repeats the steps 2 to 4 for F-->B and F<--B.
9. Waits for all its send requests to complete. We probably don't want
the function to exit before delivering data, although it may still work
and we may can gain some performance by not waiting before starting
communicating velocities. I haven't tried that, but even if it works, it
may be a bit dangerous.

No barrier seems to be required. A rank can change the dimension it
communicates without worrying what is happening far away from itself. It
only needs to be ready on a local level.

Added some example output to illustrate the difference between blocking
and non-blocking communication. In this, an earlier implementation is
used, but it is still useful.

A directory misc/ was created, to put there non-code-related files, like
the gprof output. The debug_output.txt was not needed anymore.