Implemented "increase checkpoint interval".
The user can now specify that the number of iterations between
checkpoints should increase as the simulation proceeds. New form of the
checkpoint options node:

<checkpoint iterations="10" increaseIter="true" maxIter="20"
incrFactor="1.2" cleanDirectory="true">

All the new tags (increaseIter, maxIter, incrFactor) are optional.
- increaseIter defaults to "false".
- maxIter defaults to -1, that mean "no limit".
- incrFactor defaults to 1.1 (10% increase every time).

Every time a checkpoint is printed, the parameters.checkpoint.iterations
gets increased by "incrFactor", starting with "iterations", with the
option to be bounded by "maxIter". For example, the above config creates
checkpoints for the following timesteps:
10, 22, 36, 52, 71, 91, 111, 131, 151 (max), 171, 191, 211, 226 (last).